Replace HTTPotion with httpc
The current requirements for posting to google spreadsheet are very
basic, and so using a full blown http client such as HTTPotion is
perhaps overkill. This commit refactors the code to use the existing
httpc library that ships with erlang/otp.

A new module called Http is added which wraps the httpc stuff, mainly
because when dealing with any erlang library, we need to switch between
charlists and strings. Also, our current test assumes the response is a
struct, so we can write `response.status_code`. For that reason, our new
Http module responds with a %Http.Response{} struct, with a minimal set
of fields defined.

If more functionality is required in the future, we can add it, but for
now, this seems like the bare minimum required to achieve the desired
outcome.

Happy to discuss this.